Push du 30/12
Lire le fichier CHANGEMENTS.md pour avoir le détail des modifications

+++ src/main/java/aubay/lu/projetrh/service/implementation/UtilisateurServiceImpl.java
@@ -308,7 +308,9 @@ public class UtilisateurServiceImpl implements UtilisateurService {
     public void deleteCandidat(UUID candidatId) {
         Optional<Utilisateur> candidatToDelete = utilisateurRepository.findById(candidatId);
         if(candidatToDelete.isPresent()){
-            utilisateurRepository.deleteById(candidatId);
+            if(candidatToDelete.get().getRole().getId() == 3){
+                utilisateurRepository.deleteById(candidatId);
+            }
         }
     }
 }
